+++ common/src/main/java/net/mehvahdjukaar/snowyspirit/SnowySpirit.java
@@ -4,6 +4,7 @@ package net.mehvahdjukaar.snowyspirit;
 import net.mehvahdjukaar.moonlight.api.platform.PlatformHelper;
 import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.mehvahdjukaar.snowyspirit.common.network.NetworkHandler;
+import net.mehvahdjukaar.snowyspirit.configs.ClientConfigs;
 import net.mehvahdjukaar.snowyspirit.configs.ModConfigs;
 import net.mehvahdjukaar.snowyspirit.dynamicpack.ClientDynamicResourcesHandler;
 import net.mehvahdjukaar.snowyspirit.dynamicpack.ServerDynamicResourcesHandler;
@@ -40,6 +41,7 @@ public class SnowySpirit {
 
     public static void commonInit() {
         ModConfigs.init();
+        ClientConfigs.init();
 
         NetworkHandler.init();