+++ forge/src/main/java/net/mehvahdjukaar/snowyspirit/forge/SnowySpiritForgeClient.java
@@ -1,18 +1,9 @@
 package net.mehvahdjukaar.snowyspirit.forge;
 
 import net.mehvahdjukaar.snowyspirit.SnowySpirit;
-import net.mehvahdjukaar.snowyspirit.client.SammysParticleHacks;
 import net.mehvahdjukaar.snowyspirit.integration.configured.ModConfigSelectScreen;
 import net.mehvahdjukaar.snowyspirit.reg.ClientRegistry;
-import net.minecraft.world.entity.animal.Sheep;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.ShearsItem;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.RenderLevelLastEvent;
-import net.minecraftforge.client.event.RenderLevelStageEvent;
-import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
@@ -34,13 +25,4 @@ public class SnowySpiritForgeClient {
         });
     }
 
-//    @Mod.EventBusSubscriber(modid = SnowySpirit.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.FORGE)
-//    public static class ClientEvents {
-//        @SubscribeEvent
-//        public static void onRenderStage(RenderLevelStageEvent event) {
-//            if ((event.getStage() == RenderLevelStageEvent.Stage.AFTER_TRIPWIRE_BLOCKS)) {
-//           //     SammysParticleHacks.renderLast(event.getPoseStack());
-//            }
-//        }
-//    }
 }