+++ forge/src/main/java/net/mehvahdjukaar/snowyspirit/integration/configured/ModConfigScreen.java
@@ -1,7 +1,7 @@
 package net.mehvahdjukaar.snowyspirit.integration.configured;
 
 
-import com.mojang.blaze3d.vertex.*;
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.mrcrayfish.configured.api.IModConfig;
 import com.mrcrayfish.configured.client.util.ScreenUtil;
 import net.mehvahdjukaar.moonlight.api.integration.configured.CustomConfigScreen;
@@ -34,6 +34,9 @@ public class ModConfigScreen extends CustomConfigScreen {
         addIcon("blocks and items", ModRegistry.CANDY_CANE_BLOCK.get());
         addIcon("snowy season", ModRegistry.SNOW_GLOBE.get());
         addIcon("misc", ModRegistry.GINGERBREAD_COOKIE.get());
+        addIcon("friction", Items.SNOW_BLOCK);
+        addIcon("steering", Items.LEAD);
+        addIcon("particles", ModRegistry.GLOW_LIGHTS_ITEMS.get(null).get().asItem());
     }
 
     public ModConfigScreen(CustomConfigSelectScreen parent, IModConfig config) {
@@ -66,10 +69,10 @@ public class ModConfigScreen extends CustomConfigScreen {
         super.render(poseStack, mouseX, mouseY, partialTicks);
 
         var level = Minecraft.getInstance().level;
-        if(level != null && SnowySpirit.isChristmasSeason(level)) {
+        if (level != null && SnowySpirit.isChristmasSeason(level)) {
             int x = (int) (this.width * 0.93f);
             this.itemRenderer.renderAndDecorateFakeItem(Items.SNOWBALL.getDefaultInstance(), x, 16);
-            if (ScreenUtil.isMouseWithin(x , 16, 16, 16, mouseX, mouseY)) {
+            if (ScreenUtil.isMouseWithin(x, 16, 16, 16, mouseX, mouseY)) {
                 this.renderTooltip(poseStack, this.font.split(Component.translatable("gui.snowyspirit.snow_season_on").withStyle(ChatFormatting.AQUA), 200), mouseX, mouseY);
             }
         }