+++ changelog.md
@@ -1,3 +1,5 @@
-- Added a bunch of configs to control sleds physics
-- Decreased ginger flowers dropped by crop
-- Fixed double ginger slab model
+- Improved ginger loot table
+- Fixed missing glow light particle texture
+- Fixed gingerbread vertical slab model
+- Fixed some incorrect configs
+- Added some client configs to alter glow particles. These give you the choice to choose between 3 render modes